requirements.txt
botbuilder-integration-aiohttp>=4.14.0
psycopg2
langchain
langchain_community
langchain-google-genai
google.genrativeai